wip: marked getValue as deprecated
