+++ get-value/index.js
@@ -1,12 +1,3 @@
 export function getValue(store) {
-  if (store.active) {
-    return store.value
-  } else {
-    let result
-    let unbind = store.subscribe(value => {
-      result = value
-    })
-    unbind()
-    return result
-  }
+  return store.get()
 }
