+++ index.d.ts
@@ -4,12 +4,7 @@ export {
   BuilderStore,
   MapBuilder
 } from './map-template/index.js'
-export {
-  ReadableStore,
-  WritableStore,
-  atom,
-  StoreValue
-} from './atom/index.js'
+export { ReadableStore, WritableStore, atom, StoreValue } from './atom/index.js'
 export {
   clearEffects,
   startEffect,
@@ -21,7 +16,11 @@ export { clean, cleanStores } from './clean-stores/index.js'
 export { update, updateKey } from './update/index.js'
 export { computed } from './computed/index.js'
 export { keepActive } from './keep-active/index.js'
-export { getValue } from './get-value/index.js'
 export { mount, STORE_CLEAN_DELAY } from './mount/index.js'
 
-export { createStore, createDerived, defineMap } from './deprecated/index.js'
+export {
+  createStore,
+  createDerived,
+  defineMap,
+  getValue
+} from './deprecated/index.js'
