+++ index.js
@@ -11,12 +11,12 @@ export { update, updateKey } from './update/index.js'
 export { computed } from './computed/index.js'
 export { keepActive } from './keep-active/index.js'
 export { map } from './map/index.js'
-export { getValue } from './get-value/index.js'
 export { mount, STORE_CLEAN_DELAY } from './mount/index.js'
 
 export {
   createStore,
   createDerived,
   defineMap,
-  createMap
+  createMap,
+  getValue
 } from './deprecated/index.js'
